remove build tags input for simplicity
There's no need to pass build tags as an input through the action, we
can instead pass the command to execute through to the container which
will make this action easier to maintain in the long run.  We simply
pass the command along and otherwise get out of the way.