simplify arguments for docker run
We don't want to be in the business of parsing env vars for Docker.
Simplify by allowing the caller to pass arbitrary arguments to the
docker run command.

The use case for private registries is to use the step env field to set
the environment, then pass `-e FOO` to docker to pass the env var
through to the container context.